handler package (token from access key, add quantiles, detect region, rw retry, logging)
