+++ handler/handler_test.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"context"
@@ -11,8 +11,8 @@ import (
 	"testing"
 )
 
-func TestMain(m *testing.M) {
-	jsonFile, err := os.Open("testEvent.json")
+func TestHandleRequest(t *testing.T) {
+	jsonFile, err := os.Open("../testdata/validEvent.json")
 	// if we os.Open returns an error then handle it
 	if err != nil {
 		fmt.Println(err)
@@ -22,6 +22,6 @@ func TestMain(m *testing.M) {
 	byteValue, _ := ioutil.ReadAll(jsonFile)
 	request := events.APIGatewayProxyRequest{}
 	json.Unmarshal(byteValue, &request)
-	result, _ := handleRequest(ctx, request)
+	result, _ := HandleRequest(ctx, request)
 	log.Println(result)
 }
